+++ app/(route)/(auth)/login/_components/log-in-form.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import type { FieldError, RegisterOptions } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
-import type { LoginFormValues } from '@/_apis/type';
+import type { LoginFormValues } from '@/api/type';
 import useUserStore from '@/store/useUserStore';
 
 import { useModal } from '@/_hooks/useModal';
