+++ app/(route)/(auth)/signup/_components/sign-up-form.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import type { FieldError, RegisterOptions } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
-import type { SignupFormValues } from '@/_apis/type';
-import { postSignup } from '@/_apis/user';
+import type { SignupFormValues } from '@/api/type';
+import { postSignup } from '@/api/user';
 
 import { useModal } from '@/_hooks/useModal';
