+++ app/(route)/activity/register/_components/FileInput/index.tsx
@@ -2,7 +2,7 @@
 
 import { forwardRef } from 'react';
 import Image from 'next/image';
-import type { SubImage } from '@/_apis/activities/getActivity';
+import type { SubImage } from '@/api/activities/getActivity';
 
 import DeleteIcon from 'public/assets/icons/delete.svg';
 import PlusIcon from 'public/assets/icons/plus.svg';
@@ -22,27 +22,27 @@ interface FileInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
 export default forwardRef(function FileInput({ images, editImages, onClear, ...rest }: FileInputProps, ref: React.LegacyRef<HTMLInputElement>) {
   return (
     <div className="grid">
-      <div className="pc:gap-6 flex flex-wrap items-center gap-2 tablet:gap-4">
-        <div className="pc:w-[calc((100%-72px)/4)] relative flex aspect-square w-[calc((100%-8px)/2)] flex-col items-center justify-center gap-[30px] rounded-xl border border-dashed border-gray-700 p-2 tablet:w-[calc((100%-16px)/2)]">
+      <div className="flex flex-wrap items-center gap-2 tablet:gap-4 pc:gap-6">
+        <div className="relative flex aspect-square w-[calc((100%-8px)/2)] flex-col items-center justify-center gap-[30px] rounded-xl border border-dashed border-gray-700 p-2 tablet:w-[calc((100%-16px)/2)] pc:w-[calc((100%-72px)/4)]">
           <Image src={PlusIcon} alt="등록" />
           <span className="text-center text-xl leading-[1.1] text-gray-700">이미지 등록</span>
           <input type="file" ref={ref} {...rest} className="absolute left-0 top-0 size-full cursor-pointer opacity-0" />
         </div>
         {images?.map((image, idx) => (
-          <div key={idx} className="pc:w-[calc((100%-72px)/4)] relative aspect-square w-[calc((100%-8px)/2)] rounded-xl tablet:w-[calc((100%-16px)/2)]">
+          <div key={idx} className="relative aspect-square w-[calc((100%-8px)/2)] rounded-xl tablet:w-[calc((100%-16px)/2)] pc:w-[calc((100%-72px)/4)]">
             <Image fill sizes="max-width:100%" src={image} alt="이미지 미리보기" priority style={{ objectFit: 'contain', borderRadius: '12px' }} />
-            <button type="button" onClick={() => onClear(image)} className="pc:-right-5 pc:-top-5 pc:size-10 absolute -right-2 -top-2 size-6">
+            <button type="button" onClick={() => onClear(image)} className="absolute -right-2 -top-2 size-6 pc:-right-5 pc:-top-5 pc:size-10">
               <Image src={DeleteIcon} alt="삭제" />
             </button>
           </div>
         ))}
         {editImages?.map((image, idx) => (
-          <div key={idx} className="pc:w-[calc((100%-72px)/4)] relative aspect-square w-[calc((100%-8px)/2)] rounded-xl tablet:w-[calc((100%-16px)/2)]">
+          <div key={idx} className="relative aspect-square w-[calc((100%-8px)/2)] rounded-xl tablet:w-[calc((100%-16px)/2)] pc:w-[calc((100%-72px)/4)]">
             <Image fill sizes="max-width:100%" src={image.imageUrl} alt="이미지 미리보기" priority style={{ objectFit: 'contain', borderRadius: '12px' }} />
             <button
               type="button"
               onClick={() => onClear(image.imageUrl, image?.id)}
-              className="pc:-right-5 pc:-top-5 pc:size-10 absolute -right-2 -top-2 size-6"
+              className="absolute -right-2 -top-2 size-6 pc:-right-5 pc:-top-5 pc:size-10"
             >
               <Image src={DeleteIcon} alt="삭제" />
             </button>
