+++ app/_components/header/index.tsx
@@ -3,7 +3,7 @@
 import { useEffect } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
-import type { Response } from '@/_apis/type';
+import type { Response } from '@/api/type';
 import useUserStore from '@/store/useUserStore';
 
 import GuestHeader from './_component/guest-header';
