+++ app/api/activities/patchActivity.ts
