+++ app/api/activities/postImage.ts
